BUGFIX/MEDIUM(keepalived): Ensure service is restarted properly
 ##### SUMMARY
It appears the service reload handler is not always run, so use a
registered variable on config file templating and a conditionnal task to
do the same as the handler.

Note: `service: state: reloaded` will start the service if it wasn't, it
is a documented stable ansible module behavior.

Fixes: https://openio.atlassian.net/browse/OS-330

 ##### ISSUE TYPE
- Bugfix Pull Request

 ##### SCOPE (skeleton only)
- SDS

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION